moves participants to component and mixamo rig to a util component
