+++ admin/three-object-viewer-settings/init.php
@@ -38,7 +38,7 @@ add_action('rest_api_init', function (){
 				return rest_ensure_response( [
 					'enabled' => get_option( '3ov_ai_enabled', false ),
 					'networkWorker' => get_option( '3ov_mp_networkWorker', '' ),
-                    'multiplayerWorker' => get_option( '3ov_mp_multiplayerWorker', '' ),
+					'multiplayerWorker' => get_option( '3ov_mp_multiplayerWorker', '' ),
 					'openApiKey' => three_decrypt ( get_option( '3ov_ai_openApiKey', '' ) ),
 					'allowPublicAI' => get_option( '3ov_ai_allow', '' ),
 					'defaultVRM' => get_option( '3ov_defaultVRM', '' ),
@@ -56,7 +56,7 @@ add_action('rest_api_init', function (){
 				$data = $request->get_json_params();
 				update_option( '3ov_ai_enabled', $data['enabled'] );
 				update_option( '3ov_mp_networkWorker', $data['networkWorker'] );
-                update_option( '3ov_mp_multiplayerWorker', $data['multiplayerWorker'] );
+				update_option( '3ov_mp_multiplayerWorker', $data['multiplayerWorker'] );
 				update_option( '3ov_defaultVRM', $data['defaultVRM'] );
 				update_option( '3ov_defaultAvatar', $data['defaultAvatar'] );
 				update_option( '3ov_ai_allow', $data['allowPublicAI'] );
